requirements.txt
# Add gunicorn for production deployment
gunicorn==20.1.0

# Core dependencies
Flask==2.2.5
google-analytics-data==0.18.0
openai==1.3.0
# Optional: for Gemini as LLM provider
google-generativeai>=0.8.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.2.3
google-auth==2.23.0
httpx==0.25.0
starlette==0.27.0
uvicorn==0.24.0

# Google Cloud Storage for report storage
google-cloud-storage==2.10.0

# Google Secret Manager (optional: when SECRET_MANAGER=true in standalone)
google-cloud-secret-manager==2.16.0

# Web scraping for page analysis
beautifulsoup4==4.12.2 